requirements.txt
faster-whisper>=0.10.0
torch==2.1.0
torchaudio==2.1.0
pydub>=0.25.1
numpy<2.0.0
psutil>=5.9.0
tqdm>=4.65.0
ffmpeg-python>=0.2.0 